Create Project 1
This is a program that I've been working on for about a month and a half. the reason this is all i have after a good month and a half is because I started programming in python a month and a half ago. So any pointers would be great help!